refactor: extract readiness-core package to deduplicate audit logic
