+++ tools/goal-audit/package.json
@@ -45,5 +45,8 @@
   "devDependencies": {
     "@types/node": "^25.9.3",
     "typescript": "^5.0.0"
+  },
+  "dependencies": {
+    "@cobusgreyling/readiness-core": "file:../readiness-core"
   }
 }
