+++ tools/goal-audit/src/auditor.ts
@@ -1,5 +1,6 @@
-import { readdir, readFile, stat } from 'node:fs/promises';
+import { readFile } from 'node:fs/promises';
 import path from 'node:path';
+import { Finding, BaseAuditResult, fileExists, scanSkillDirectories } from '@cobusgreyling/readiness-core';
 
 export interface GoalSignals {
   goalFile: { present: boolean; path?: string };
@@ -15,20 +16,9 @@ export interface GoalSignals {
   runLog: { present: boolean };
 }
 
-export interface Finding {
-  level: 'ok' | 'warn' | 'fail';
-  message: string;
-}
+export type { Finding };
 
-export interface AuditResult {
-  target: string;
-  score: number;
-  level: 'G0' | 'G1' | 'G2' | 'G3';
-  assessment: string;
-  signals: GoalSignals;
-  findings: Finding[];
-  recommendations: string[];
-}
+export interface AuditResult extends BaseAuditResult<'G0' | 'G1' | 'G2' | 'G3', GoalSignals> {}
 
 const GOAL_FILES = ['GOAL.md', 'goal.md', 'docs/GOAL.md'];
 const GOAL_SKILL_NAMES = [
@@ -40,31 +30,8 @@ const SAFETY_FILES = ['docs/safety.md', 'SECURITY.md', 'safety.md'];
 const BUDGET_FILES = ['goal-budget.md', 'docs/goal-budget.md'];
 const RUN_LOG_FILES = ['goal-run-log.md', 'docs/goal-run-log.md'];
 
-async function fileExists(p: string): Promise<boolean> {
-  try {
-    await stat(p);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
 async function findSkills(root: string): Promise<string[]> {
-  const dirs = [
-    path.join(root, '.grok', 'skills'),
-    path.join(root, '.claude', 'skills'),
-    path.join(root, '.codex', 'skills'),
-    path.join(root, 'skills'),
-  ];
-  const found: string[] = [];
-  for (const dir of dirs) {
-    if (!(await fileExists(dir))) continue;
-    const entries = await readdir(dir, { withFileTypes: true });
-    for (const e of entries) {
-      if (e.isDirectory()) found.push(e.name);
-    }
-  }
-  return found;
+  return scanSkillDirectories(root);
 }
 
 async function detectTests(root: string): Promise<boolean> {
