+++ tools/loop-audit/package.json
@@ -50,5 +50,8 @@
   "devDependencies": {
     "@types/node": "^26.0.0",
     "typescript": "^7.0.2"
+  },
+  "dependencies": {
+    "@cobusgreyling/readiness-core": "file:../readiness-core"
   }
 }
